refactor(model): divide Result class
